refactor(canvas): co-locate share-link parser with its builders
Move parseShareLink next to canvasShareUrl/channelShareUrl in posthogLinks.ts
so the /code/canvas and /code/channel path shapes live in one place (the
parser is the inverse of the builders). Derive the recognized-host set from
REGION_LABELS instead of a hardcoded region list. shareLinks.ts is now just
the navigation glue (navigateToShareTarget/handleShareLinkClick).

Generated-By: PostHog Code
Task-Id: a3001315-524a-4902-b23c-5c6f5b010827
