+++ packages/ui/src/utils/posthogLinks.ts
@@ -1,4 +1,8 @@
-import type { CloudRegion } from "@posthog/shared";
+import {
+  type CloudRegion,
+  getCloudUrlFromRegion,
+  REGION_LABELS,
+} from "@posthog/shared";
 import { useAuthStore } from "@posthog/ui/features/auth/store";
 import { getPostHogUrl } from "@posthog/ui/utils/urls";
 
@@ -115,6 +119,76 @@ export function channelShareUrl(
   return getPostHogUrl(
     taskId ? `${base}/tasks/${encodeURIComponent(taskId)}` : base,
   );
+}
+
+/**
+ * The in-app destination a PostHog Code share link points at — the inverse of
+ * the `canvasShareUrl` / `channelShareUrl` builders above.
+ */
+export type ShareLinkTarget =
+  | { kind: "canvas"; channelId: string; dashboardId: string }
+  | { kind: "channel"; channelId: string; taskId?: string };
+
+// Hosts we recognise as PostHog share-link origins, one per cloud region. The
+// host check keeps `parseShareLink` from hijacking unrelated links that happen
+// to share the `/code/...` path shape.
+const POSTHOG_HOSTS = new Set(
+  (Object.keys(REGION_LABELS) as CloudRegion[])
+    .map((region) => {
+      try {
+        return new URL(getCloudUrlFromRegion(region)).host;
+      } catch {
+        return "";
+      }
+    })
+    .filter(Boolean),
+);
+
+/**
+ * Parse a PostHog Code share link into its in-app navigation target, or `null`
+ * if it isn't one. Recognises the `/code/canvas/...` and `/code/channel/...`
+ * links built above, on any region's host — the inbound deep-link handlers
+ * navigate by id against the current session, so bouncing through the browser
+ * to reach the app buys nothing.
+ */
+export function parseShareLink(href: string): ShareLinkTarget | null {
+  let url: URL;
+  try {
+    url = new URL(href);
+  } catch {
+    return null;
+  }
+  if (!POSTHOG_HOSTS.has(url.host)) return null;
+
+  // Split the still-encoded pathname first, then decode each segment, so an id
+  // containing an encoded slash (`%2F`) stays a single segment.
+  const segments = url.pathname
+    .split("/")
+    .filter(Boolean)
+    .map((segment) => {
+      try {
+        return decodeURIComponent(segment);
+      } catch {
+        return segment;
+      }
+    });
+
+  if (segments[0] !== "code") return null;
+
+  if (segments[1] === "canvas" && segments.length === 4) {
+    return { kind: "canvas", channelId: segments[2], dashboardId: segments[3] };
+  }
+
+  if (segments[1] === "channel") {
+    if (segments.length === 3) {
+      return { kind: "channel", channelId: segments[2] };
+    }
+    if (segments.length === 5 && segments[3] === "tasks") {
+      return { kind: "channel", channelId: segments[2], taskId: segments[4] };
+    }
+  }
+
+  return null;
 }
 
 export function errorTrackingIssueUrl(
